cmake: ターゲット INSTALL の修正 (find_package 対応)

+++ core/src/HumanInterface/DRUartMotorDriver.h
@@ -55,7 +55,7 @@ public:
 	///	コンストラクタ	chは背面のスイッチになる予定
 	DRUARTMotorDriver(const DRUARTMotorDriverDesc& d = DRUARTMotorDriverDesc());
 	virtual ~DRUARTMotorDriver();
-	int NMotor() { return counts.size(); }
+	int NMotor() { return (int) counts.size(); }
 
 	///	初期化
 	virtual bool Init();
